Square into squares

+++ src/sum_by_factors.py
@@ -25,13 +25,13 @@ def sum_for_list(lst):
 def find_prime_factors(num):
     """Finds all the prime factors of a specific integer"""
     factors = []
-    for pos in range(1, math.floor(math.fabs(num) ** .5) + 1):
+    for pos in range(1, int(math.floor(math.fabs(num) ** .5) + 1)):
         if num % pos == 0:
             factors.append(pos)
             factors.append(math.fabs(int(num / pos)))
     for ind in range(len(factors) - 1, 0, -1):
         factor = factors[ind]
-        for pos in range(2, math.floor(factor ** .5) + 1):
+        for pos in range(2, int(math.floor(factor ** .5) + 1)):
             if factor % pos == 0:
                 factors.pop(factors.index(factor))
                 break
